+++ common/video/iso-mpeg4/src/sys_encoder_shpenc.cpp
@@ -902,9 +902,9 @@ UInt CVideoObjectEncoder::encodeSIIntra (ShapeSSMode shpmd, Bool* HorSamplingChk
 	PixelC* ppxlcReconCurrBABTr = NULL;	//for st_order
 	Int scan_order =0;					
 
-	double h_factor = log((double)m_volmd.ihor_sampling_factor_n_shape/m_volmd.ihor_sampling_factor_m_shape)/log(2);
+	double h_factor = log((double)m_volmd.ihor_sampling_factor_n_shape/m_volmd.ihor_sampling_factor_m_shape)/log((double)2);
 	int h_factor_int = (int)floor(h_factor+0.000001);
-	double v_factor = log((double)m_volmd.iver_sampling_factor_n_shape/m_volmd.iver_sampling_factor_m_shape)/log(2);
+	double v_factor = log((double)m_volmd.iver_sampling_factor_n_shape/m_volmd.iver_sampling_factor_m_shape)/log((double)2);
 	int v_factor_int = (int)floor(v_factor+0.000001);
 
 	Int NumTwoPowerLoopX = h_factor_int;
